Move HttpsRedirectService into its own file
