+++ src/main.rs
@@ -13,6 +13,7 @@ mod error;
 mod file_stream;
 mod files;
 mod html;
+mod https_redirect_service;
 mod input_path;
 mod redirect;
 mod request_handler;
